Add elf txn flow test case in shim test
